tests/wpt/web-platform-tests/XMLHttpRequest/getallresponseheaders.htm

<!doctype html>
<title>XMLHttpRequest: getAllResponseHeaders()</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<div id="log"></div>
<script>
async_test((t) => {
  const client = new XMLHttpRequest()
  client.onload = t.step_func_done(() => {
    assert_equals(client.getAllResponseHeaders(), "foo-test: 1, 2, 3\r\n")
  })
  client.open("GET", "resources/headers-basic.asis")
  client.send(null)
})

async_test((t) => {
  const client = new XMLHttpRequest()
  client.onload = t.step_func_done(() => {
    assert_equals(client.getAllResponseHeaders(), "also-here: Mr. PB\r\newok: lego\r\nfoo-test: 1, 2\r\n")
  })
  client.open("GET", "resources/headers.asis")
  client.send(null)
})
</script>
